fix: sonarcloud feedback duplication code ping 3826
